Cutting it down to two pages
For the time being I'm doing it wholesale, cutting whole paragraphs and eliminating figures. We'll have to refine a bit more. Refs #2

Please @mariosky and @ccottap take a look and cut down non-coherent parts to leave the core, which is basically what we're doing now and asking for help on how to do it better.